pyproject.toml
[tool.pyright]
exclude = [
  "**/node_modules",
  "**/__pycache__",
]
venvPath = "venvs"
venv = "debbugs-summarize"

[project]
name = "debbugs-summarize"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "google-generativeai",
]
authors = [
  {name = "dickmao", email = "richard@commandlinesystems.com"},
]

[project.optional-dependencies]
test = [
  "pytest",
]
